StateHandlerを削除
